+++ docs/CHANGELOG.rst
@@ -25,6 +25,8 @@ Changed
   - Old: Registry, Monitor and System Controller prompt the user by default to press `[Enter]`` to end process after shutdown.
   - New: All utilities terminate without prompting for user input. SystemController option '-ni, --non-interactive' is deprecated.
 
+- Implemented the union (de-)serialization stubs in the ``silkit/util/serdes`` headers.
+
 
 [4.0.52] - 2024-09-02 
 ---------------------